InMemory locking overhaul
No more maintain_lock() and refresh_lock().

Instead locks last forever, but are auto-dropped if the pid that locked
it is no longer alive.

You can still specify an unlock_after for use as a timer.

This means we no longer fork to maintain a refresher and no longer have
to wait until default timer runs out before getting a lock, but it does
mean we'll do a lot of ssh calls...