update bam_processing step for cram input and output and indexing
the submissions for bam_processing are now with a dispatch_vrpipecode call
rather than dispatch call, so one needs to be a little careful about escaping
characters in the ad hoc command lines. e.g. the awk script in the
`chipseq_qc_and_peak_calling.t` test was modified in this commit escape the
`$` characters