Adds --max_processes cmdline arg and sane defaults
Previously the maximum number of simultaneous forks
was set to the number of cores in the machine. That
is insanely low in some cases (e.g. on a single core
machine the http server becomes single threaded and
thus cannot handle more than one API call at a time
without blocking).