required reads metadata for bwa_mem_align_unitigs
when the fermikit pipeline was reset, can have a different number
of reads in the assembled unitigs as it does not give the exact
same results each time. this means the read metadata in the
aligned bam from a previous run was being used as the expected
read count, but this was out-of-date. now add the metadata in
the fermi2_simplify step and pass it through